feat(comparers)!: refactor ArrayEqualityComparer and ArrayEqualityComparerTests
- move `ArrayEqualityComparer` to `Comparers` namespace
- update `ArrayEqualityComparerTests` to use `TheoryData` instead of `IEnumerable<object[]>`
- remove unnecessary using directives in `ArrayEqualityComparerTests`
